requirements.txt
Flask==3.0.3
ImageHash==4.3.1
mlx_whisper==0.3.0
openai==1.46.0
numpy==1.24.3
opencv_python==4.8.1.78
opencv_python_headless==4.10.0.84
Pillow==10.4.0
yt_dlp==2024.8.6